Split GC type layout computation into a shared trait in `wasmtime_environ`
This stuff was previously living in the `GcRuntime` trait, but it turns out that the
compiler will also need to know the layout of GC types, who'd have thunk it.
