+++ crates/wasmtime/src/runtime/vm/gc/enabled/arrayref.rs
@@ -3,11 +3,10 @@ use crate::{
     prelude::*,
     runtime::vm::{GcHeap, GcStore, VMGcRef},
     store::{AutoAssertNoGc, StoreOpaque},
-    vm::GcArrayLayout,
     AnyRef, ExternRef, HeapType, RootedGcRefImpl, StorageType, Val, ValType,
 };
 use core::fmt;
-use wasmtime_environ::VMGcKind;
+use wasmtime_environ::{GcArrayLayout, VMGcKind};
 
 /// A `VMGcRef` that we know points to a `array`.
 ///
